+++ CHANGELOG.md
@@ -8,8 +8,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [UNRELEASED]
 
 ### Added
+  * for Arbiter
+    * BusinessApplication (ba_) scope as a CheckDigitScope
+    * PendingEvent (pe_) scope as a CheckDigitScope
+    * DisposedEvent (de_) scope as a CheckDigitScope
+    * AuditEntry (ae_)  scope as a CheckDigitScope
 ### Removed
+  * for Arbiter
+    * Activity (act_) scope
+    * ExternalEvent(ee_) scope
+    * BusinessApplication (be_) scope
+  * Removed  enum NoCheckdigitArbiterScopes, as fully deprecated
+  * Removed tests for the NoCheckdigitArbiterScopes
 ### Changed
+  * for Arbiter 
+    * changed Workflow (wf_) scope to CheckDigitScope
   * Removed `approve_for_publish` step from master:build
 ### Fixed
 ### Deprecated